refactor: simplify firebase
