--- src/utils/services/firebase/index.ts
@@ -1,3 +1,0 @@
-import './init'
-
-export * from './storage'
