--- src/utils/services/firebase/init.ts
@@ -1,10 +1,0 @@
-import { initializeApp } from 'firebase/app'
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-}
-
-initializeApp(firebaseConfig)
